+++ CHANGELOG.md
@@ -7,8 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - API diagram in user guide
-- Generic `Metadata` system for BayBE objects containing description, unit, and
-  additional information
+- `Metadata` class providing description, unit, and other information for BayBE objects
 
 ### Fixed
 - `Campaign` no longer allows overlapping names between parameters and targets